apps/user_auth/views.py
from django.shortcuts import render
from rest_framework import permissions, viewsets
from rest_framework.response import Response
from rest_framework.views import APIView
from apps.user_auth.models import RecipeJarUser
from apps.user_auth.serializer import UserSerializer
from rest_framework import status

