requirements.txt
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
beautifulsoup4==4.12.2
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
distro==1.8.0
dj-database-url==2.1.0
Django==4.2.8
djangorestframework==3.14.0
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.0
extruct==0.16.0
google-api-core==2.15.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
googleapis-common-protos==1.62.0
gunicorn==21.2.0
h11==0.14.0
html-text==0.5.2
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.6
inflect==7.0.0
ingredient-parser==1.0.2
isodate==0.6.1
Jinja2==3.1.2
jstyleson==0.0.2
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
mf2py==2.0.1
murmurhash==1.0.10
num2words==0.5.13
numpy==1.26.2
openai==1.3.8
packaging==23.2
preshed==3.0.9
protobuf==4.25.1
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
pyRdfa3==3.5.3
pytz==2023.3.post1
quantulum3==0.9.0
rdflib==7.0.0
recipe-scrapers==14.55.0
requests==2.31.0
rsa==4.9
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
w3lib==2.1.2
wasabi==1.1.2
weasel==0.3.4
webencodings==0.5.1
measurement==4.0a8